feat: Update service/controller implemented, fix: missing data in db issue

+++ src/main/java/com/jr/tdl/main/persistence/repository/UserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.jr.tdl.main.persistence.entity.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Float>{
+public interface UserRepository extends JpaRepository<User, Long>{
 	
 
 }
